# perf: don't create unnecessary lock with Context._with_span usage