--- a/ddtrace/context.py
+++ b/ddtrace/context.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Any
 from typing import Optional
 from typing import TYPE_CHECKING
@@ -6,6 +5,7 @@ from typing import Text
 
 from .constants import ORIGIN_KEY
 from .constants import SAMPLING_PRIORITY_KEY
+from .internal import forksafe
 from .internal.compat import NumericType
 from .internal.logger import get_logger
 from .internal.utils.deprecation import deprecated
@@ -40,7 +40,7 @@ class Context(object):
         sampling_priority=None,  # type: Optional[float]
         meta=None,  # type: Optional[_MetaDictType]
         metrics=None,  # type: Optional[_MetricDictType]
-        lock=None,  # type: Optional[threading.RLock]
+        lock=None,  # type: Optional[forksafe.RLock]
     ):
         self._meta = meta if meta is not None else {}  # type: _MetaDictType
         self._metrics = metrics if metrics is not None else {}  # type: _MetricDictType
@@ -56,7 +56,7 @@ class Context(object):
         if lock is not None:
             self._lock = lock
         else:
-            self._lock = threading.RLock()
+            self._lock = forksafe.RLock()
 
     def _with_span(self, span):
         # type: (Span) -> Context