## Commit Message
<!-- Defaults to the title of the PR. Replace if desired. -->
{{title}}

Small optimization to avoid creating a `threading.RLock` or two dicts if we are just going to overwrite them anyways.

This gives us a very tiny performance optimization, but mostly helps us avoid creating an additional RLock/two dicts which will save us on object creation + GC.


**Benchmark: tracer**
| Benchmark      | _master_ | _this branch_ |
| -- | -- | -- |
| tracer-small | 506 us | 500 us: 1.01x faster | 
| tracer-medium  | 4.90 ms | 4.79 ms: 1.02x faster |
| tracer-large   | 50.6 ms | 47.2 ms: 1.07x faster |
| Geometric mean | (ref) | 1.03x faster |

**Benchmark: flask_simple**
| Benchmark      | _master_ | _this branch_ |
| -- | -- | -- |
| flasksimple-baseline | 4.66 ms | 3.72 ms: 1.25x faster | 
| flasksimple-tracer   | 4.31 ms | 3.78 ms: 1.14x faster |
| flasksimple-profiler | 4.06 ms | 3.74 ms: 1.09x faster |
| Geometric mean | (ref) | 1.11x faster |

**Benchmark: django_simple**
No significant change.

Given the results from Flask I was expecting Django to also show an improvement.